requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pillow==10.0.0
tqdm==4.66.1
matplotlib==3.7.2
pandas==2.0.3
jupyter==1.0.0
notebook==7.0.3
pytest==7.4.0
black==23.7.0
isort==5.12.0
flake8==6.1.0 